requirements.txt
streamlit
google-generativeai
Pillow
python-dotenv
streamlit-webrtc
aiortc
opencv-python-headless
av
pypdf
python-docx